+++ src/main/java/com/technova/technov/domain/service/MensajeDirectoService.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 public interface MensajeDirectoService {
     List<MensajeDirectoDto> listarTodos();
-    List<MensajeDirectoDto> listarPorUsuario(Long userId);
-    List<MensajeDirectoDto> listarPorEmpleado(Long empleadoId);
+    List<MensajeDirectoDto> listarPorUsuario(Integer userId);
+    List<MensajeDirectoDto> listarPorEmpleado(Integer empleadoId);
     List<MensajeDirectoDto> listarPorConversacion(String conversationId);
     MensajeDirectoDto crear(MensajeDirectoDto dto);
-    MensajeDirectoDto crearConversacion(Long userId, String asunto, String mensaje, String prioridad);
-    MensajeDirectoDto responderMensaje(Long parentMessageId, Long senderId, String senderType, String mensaje);
-    MensajeDirectoDto marcarLeido(Long id);
-    MensajeDirectoDto obtenerPorId(Long id);
+    MensajeDirectoDto crearConversacion(Integer userId, String asunto, String mensaje, String prioridad);
+    MensajeDirectoDto responderMensaje(Integer parentMessageId, Integer senderId, String senderType, String mensaje);
+    MensajeDirectoDto marcarLeido(Integer id);
+    MensajeDirectoDto obtenerPorId(Integer id);
 }
